catalyst-macros: support use from outside of the find_package scope
VTK and ParaView perform `find_package` in scopes outside of the
intended uses in order to avoid complex conditionals to know whether a
package is wanted at the top-level. In this case, the
`catalyst::catalyst` target is not available (without making it
`GLOBAL`), so allow callers to use their own target name if they have
one.